# Swagger/documentation.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api animalerie",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "sandbox-server"
        }
    ],
    "paths": {
...
        "/animal": {
            "get": {
                "tags": [
                    "Animal"
                ],
                "summary": "List all animals",
                "parameters": [
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List all animals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Animal[]",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Animal"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès à la page refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animal/{id}": {
            "get": {
                "tags": [
                    "Animal"
                ],
                "summary": "Retourne un animal",
                "description": "Retourne un animal",
                "operationId": "e84f7fbb9a58ad9ffc06280fd53931a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de l'animal à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne un animal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Animal"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès à la page refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animal/create": {
            "post": {
                "tags": [
                    "Animal"
                ],
                "summary": "Create animal",
                "description": "Créer un animal",
                "operationId": "6656c4c0539c20b0329a777ba83e11db",
                "parameters": [
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Animal à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "medeor"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "chien"
                                    },
                                    "race": {
                                        "type": "string",
                                        "example": "labrador"
                                    },
                                    "poids": {
                                        "type": "float",
                                        "example": 29.95
                                    },
                                    "age": {
                                        "type": "int",
                                        "example": 4
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Il est gentil"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "219779417329"
                                    },
                                    "proprietaire": {
                                        "type": "int",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Animal enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animal/update/{id}": {
            "put": {
                "tags": [
                    "Animal"
                ],
                "summary": "Update animal",
                "description": "Update animal in Db",
                "operationId": "4b531b7b2881874c27762e15858eccf0",
                "parameters": [
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de l'animal à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Animal à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "medeor"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "chien"
                                    },
                                    "race": {
                                        "type": "chien",
                                        "example": "labrador"
                                    },
                                    "poids": {
                                        "type": "float",
                                        "example": 29.95
                                    },
                                    "age": {
                                        "type": "int",
                                        "example": 4
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Il est gentil"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "219779417329"
                                    },
                                    "proprietaire": {
                                        "type": "int",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Animal enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animal/delete/{id}": {
            "delete": {
                "tags": [
                    "Animal"
                ],
                "summary": "Delete animal",
                "description": "Delete animal in Db",
                "operationId": "8c5241d414e2bf29f803d4702ee1dc93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de l'animal à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression validée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Animal supprimé"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Animal": {
                "title": "Animal",
                "properties": {
                    "id": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "nom": {
                        "nullable": false,
                        "type": "string"
                    },
                    "type": {
                        "nullable": false,
                        "type": "string"
                    },
                    "race": {
                        "nullable": false,
                        "type": "string"
                    },
                    "poids": {
                        "nullable": false,
                        "type": "float"
                    },
                    "age": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "description": {
                        "nullable": false,
                        "type": "string"
                    },
                    "photo": {
                        "nullable": false,
                        "type": "string"
                    },
                    "proprietaire": {
                        "nullable": true,
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}